requirements.txt
psycopg2-binary==2.9.11
Pillow>=10.0
python-barcode>=0.15
brother-ql>=0.9
requests==2.32.5
python-dotenv==1.2.2
